+++ app/services/reports/docx/draw_experiment.rb
@@ -22,7 +22,7 @@ module Reports::Docx::DrawExperiment
             link_style
     end
     html = custom_auto_link(experiment.description, team: @report_team)
-    html_to_word_converter(html)
+    Reports::HtmlToWordConverter.new(@docx).html_to_word_converter(html)
     @docx.p
     subject['children'].each do |child|
       public_send("draw_#{child['type_of']}", child, experiment)
