+++ app/services/reports/docx/draw_my_module.rb
@@ -66,7 +66,7 @@ module Reports::Docx::DrawMyModule
 
     if my_module.description.present?
       html = custom_auto_link(my_module.description, team: @report_team)
-      html_to_word_converter(html)
+      Reports::HtmlToWordConverter.new(@docx).html_to_word_converter(html)
     else
       @docx.p I18n.t('projects.reports.elements.module.no_description')
     end
