+++ app/services/reports/docx/draw_my_module_protocol.rb
@@ -11,7 +11,7 @@ module Reports::Docx::DrawMyModuleProtocol
                    timestamp: I18n.l(protocol.created_at, format: :full)
     @docx.hr
     html = custom_auto_link(protocol.description, team: @report_team)
-    html_to_word_converter(html)
+    Reports::HtmlToWordConverter.new(@docx).html_to_word_converter(html)
     @docx.p
     @docx.p
   end
