+++ app/services/reports/docx/draw_step_comments.rb
@@ -17,7 +17,7 @@ module Reports::Docx::DrawStepComments
                      date: I18n.l(comment_ts, format: :full_date),
                      time: I18n.l(comment_ts, format: :time)), italic: true
       html = custom_auto_link(comment.message, team: @report_team)
-      html_to_word_converter(html)
+      Reports::HtmlToWordConverter.new(@docx).html_to_word_converter(html)
       @docx.p
     end
   end
